chore: technical debt cleanup
